+++ app/microstrain_mag_cal/backend.hpp
@@ -1,16 +1,26 @@
 #pragma once
 
+#include <filesystem>
 #include <optional>
 
-#include <Eigen/Dense>
+#include <mio/mmap.hpp>
 
 #include <microstrain/array_view.hpp>
 #include <microstrain/platform.h>  // Needed to fix OS-specific naming conflicts
-#include <microstrain_mag_cal/calibration.hpp>
+#include <microstrain_mag_cal/preprocessing.hpp>
 
 namespace backend
 {
-    Eigen::MatrixX3d extractPointMatrixFromRawData(
+    struct MappedBinaryData
+    {
+        mio::mmap_source mapping;
+        microstrain::ConstU8ArrayView view;
+    };
+
+    std::optional<MappedBinaryData> mapBinaryFile(const std::filesystem::path& filepath);
+
+    microstrain_mag_cal::PointManager extractPointsFromRawData(
         const microstrain::ConstU8ArrayView &data_view,
         std::optional<double> reference_field_strength = std::nullopt);
+
 }