+++ app/microstrain_mag_cal/main.cpp
@@ -1,11 +1,10 @@
 #include <filesystem>
 #include <iostream>
 
-#include <CLI/CLI.hpp>
 #include <magic_enum/magic_enum.hpp>
-#include <mio/mmap.hpp>
 
 #include <backend.hpp>
+#include <cli.hpp>
 #include <microstrain_mag_cal/analysis.hpp>
 #include <microstrain_mag_cal/calibration.hpp>
 
@@ -13,116 +12,16 @@
 using microstrain_mag_cal::FitResult;
 
 
-// TODO: Move to view module
-// Console output after the fitting algorithms are run
-void displayFitResult(const std::string &fit_name, const FitResult &result, const double fit_RMSE)
-{
-
-    printf("\n");
-    printf("--------------------------------------------------\n");
-    printf("%s\n", fit_name.data());
-    printf("--------------------------------------------------\n");
-
-    std::string result_output = "SUCCESS";
-    if (result.error != FitResult::Error::NONE)
-    {
-        result_output = "FAIL - ";
-        const std::string_view error_name = magic_enum::enum_name(result.error);
-        result_output += error_name.empty() ? "UNKNOWN" : error_name;
-    }
-    printf("  Result: %s\n\n", result_output.c_str());
-
-    printf("  Soft-Iron Matrix:\n");
-    const Eigen::IOFormat fmt(6, 0, " ", "\n", "  ");
-    std::cout << result.soft_iron_matrix.format(fmt) << "\n\n";
-
-    printf("  Hard-Iron Offset:\n");
-    std::cout << result.hard_iron_offset.format(fmt) << "\n\n";
-
-    printf("  Fit RMSE: %.5f\n", fit_RMSE);
-}
-
-
-// TODO: Move to view module
-struct ProgramArgs
-{
-    std::filesystem::path input_data_filepath;
-
-    bool display_analysis = false;
-    bool spherical_fit = false;
-    bool ellipsoidal_fit = false;
-
-    std::optional<double> field_strength;
-
-    std::filesystem::path output_json_directory;
-};
-
-// TODO: Move to view module
-void setup_argument_parser(CLI::App& app, ProgramArgs& args, char* argv[])
-{
-    app.description("Tool to fit magnetometer calibrations for a device.");
-    app.usage("Usage: " + std::filesystem::path(argv[0]).filename().string() + " <file> [OPTIONS]");
-
-    app.add_option("file", args.input_data_filepath, "A binary file containing mip data to read from.")
-        ->check(CLI::ExistingFile)
-        ->multi_option_policy(CLI::MultiOptionPolicy::Throw)
-        ->required();
-
-    CLI::Option_group* core = app.add_option_group("Core", "Core functionality of the tool.");
-
-    core->add_flag("-a,--display-analysis", args.display_analysis, "Display comprehensive analysis output.")
-        ->multi_option_policy(CLI::MultiOptionPolicy::Throw);
-    core->add_flag("-s,--spherical-fit", args.spherical_fit, "Perform a spherical fit on the input data.")
-        ->multi_option_policy(CLI::MultiOptionPolicy::Throw);
-    core->add_flag("-e,--ellipsoidal-fit", args.ellipsoidal_fit, "Perform an ellipsoidal fit on the input data.")
-        ->multi_option_policy(CLI::MultiOptionPolicy::Throw);
-
-    core->require_option(1, 0);
-
-    app.add_option("-f,--field-strength", args.field_strength, "Field strength to use as a reference instead of using the measured field strength.")
-        ->multi_option_policy(CLI::MultiOptionPolicy::Throw);
-
-    app.add_option("-j,--output-json", args.output_json_directory, "Write the resulting calibration(s) to JSON file(s) in the given directory.")
-        ->check(CLI::ExistingDirectory)
-        ->multi_option_policy(CLI::MultiOptionPolicy::Throw);
-}
-
-// TODO: Move to backend or microstrain utilities
-struct MappedBinaryData
-{
-    mio::mmap_source mapping;
-    microstrain::ConstU8ArrayView view;
-};
-
-// TODO: Move to backend or microstrain utilities
-// TODO: Move file logic out and add test
-std::optional<MappedBinaryData> mapBinaryFile(const std::filesystem::path& filepath)
-{
-    std::error_code error;
-    mio::mmap_source mapping = mio::make_mmap_source(filepath.string(), error);
-
-    if (error)
-    {
-        return std::nullopt;
-    }
-
-    const microstrain::ConstU8ArrayView view(reinterpret_cast<const uint8_t*>(mapping.data()), mapping.size());
-
-    return MappedBinaryData{std::move(mapping), view};
-}
-
-
 int main(const int argc, char **argv)
 {
-    ProgramArgs args;
-    CLI::App app;
-    setup_argument_parser(app, args, argv);
-    CLI11_PARSE(app, argc, argv);
+    cli::ProgramArgs args(argv);
+    CLI11_PARSE(args.app, argc, argv);
 
-    const std::optional<MappedBinaryData> mapped_data = mapBinaryFile(args.input_data_filepath);
+    const std::optional<backend::MappedBinaryData> mapped_data = backend::mapBinaryFile(args.input_data_filepath);
     assert(mapped_data.has_value());
 
-    const Eigen::MatrixX3d points = backend::extractPointMatrixFromRawData(mapped_data->view, args.field_strength);
+    const microstrain_mag_cal::PointManager point_manager = backend::extractPointsFromRawData(mapped_data->view, args.field_strength);
+    const Eigen::MatrixX3d points = point_manager.getMatrix();
     const Eigen::RowVector3d initial_offset = microstrain_mag_cal::estimateInitialHardIronOffset(points);
 
     if (!args.field_strength.has_value())
@@ -138,19 +37,18 @@ int main(const int argc, char **argv)
         printf("--------------------------------------------------\n");
         printf("Analysis:\n");
         printf("--------------------------------------------------\n");
-        printf("  Used Points       : %zu\n", static_cast<size_t>(points.rows()));
+        printf("  Used Points       : %s\n", cli::getPointUsageDisplay(point_manager).c_str());
         printf("  Spatial Coverage  : %.5f%%\n", microstrain_mag_cal::calculateSpatialCoverage(points, initial_offset));
         printf("  Field Strength    : %.5f\n", args.field_strength.value());
         printf("  Initial Offset    : [%.5f, %.5f, %.5f]\n", initial_offset.x(), initial_offset.y(), initial_offset.z());
     }
 
-    // TODO: Cleanup following sections
     if (args.spherical_fit)
     {
         const FitResult fit_result = microstrain_mag_cal::fitSphere(points, args.field_strength.value(), initial_offset);
         const double fit_RMSE = microstrain_mag_cal::calculateFitRMSE(points, fit_result, args.field_strength.value());
 
-        displayFitResult("Spherical Fit", fit_result, fit_RMSE);
+        cli::displayFitResult("Spherical Fit", fit_result, fit_RMSE);
 
         if (!args.output_json_directory.empty())
         {
@@ -163,7 +61,7 @@ int main(const int argc, char **argv)
         const FitResult fit_result = microstrain_mag_cal::fitEllipsoid(points, args.field_strength.value(), initial_offset);
         const double fit_RMSE = microstrain_mag_cal::calculateFitRMSE(points, fit_result, args.field_strength.value());
 
-        displayFitResult("Ellipsoidal Fit", fit_result, fit_RMSE);
+        cli::displayFitResult("Ellipsoidal Fit", fit_result, fit_RMSE);
 
         if (!args.output_json_directory.empty())
         {